change serializer namespace
