--- app/serializers/api/base_serializer.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  class BaseSerializer
-    include FastJsonapi::ObjectSerializer
-  end
-end
