--- app/serializers/api/music_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  class MusicSerializer < ::Api::BaseSerializer
-    set_type :music
-    attributes :title, :artist
-  end
-end
