--- app/serializers/api/score_serializer.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  class ScoreSerializer < ::Api::BaseSerializer
-    set_type :score
-    attributes :user_id, :music_id, :difficulty, :points
-
-    belongs_to :user
-    belongs_to :music
-  end
-end
