--- app/serializers/api/user_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  class UserSerializer < ::Api::BaseSerializer
-    set_type :user
-    attributes :qrcode, :name
-  end
-end
