+++ app/serializers/base_serializer.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class BaseSerializer
+  include FastJsonapi::ObjectSerializer
+end
