+++ app/serializers/music_serializer.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class MusicSerializer < BaseSerializer
+  set_type :music
+  attributes :title, :artist
+end
