+++ app/serializers/user_serializer.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class UserSerializer < BaseSerializer
+  set_type :user
+  attributes :qrcode, :name
+end
